+++ src/main/java/org/polypheny/simpleclient/scenario/lockingBench/queryBuilder/CreateNamespace.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package org.polypheny.simpleclient.scenario.scalingBench.queryBuilder;
+package org.polypheny.simpleclient.scenario.lockingBench.queryBuilder;
 
 import java.util.Map;
 import kong.unirest.core.HttpRequest;
